+++ packages/@expo/cli/src/start/server/metro/MetroTerminalReporter.ts
@@ -104,8 +104,16 @@ export class MetroTerminalReporter extends TerminalReporter {
 
   _logBundlingError(error: SnippetError): void {
     const moduleResolutionError = formatUsingNodeStandardLibraryError(this.projectRoot, error);
+    const cause = error.cause as undefined | { _expoImportStack?: string };
     if (moduleResolutionError) {
-      return this.terminal.log(maybeAppendCodeFrame(moduleResolutionError, error.message));
+      let message = maybeAppendCodeFrame(moduleResolutionError, error.message);
+      if (cause?._expoImportStack) {
+        message += `\n\n${cause?._expoImportStack}`;
+      }
+      return this.terminal.log(message);
+    }
+    if (cause?._expoImportStack) {
+      error.message += `\n\n${cause._expoImportStack}`;
     }
     return super._logBundlingError(error);
   }