+++ packages/@expo/cli/src/start/server/metro/withMetroMultiPlatform.ts
@@ -109,6 +109,7 @@ export function withExtendedResolver(
   const aliases: { [key: string]: Record<string, string> } = {
     web: {
       'react-native': 'react-native-web',
+      'react-native/index': 'react-native-web',
     },
   };
 
@@ -227,9 +228,7 @@ export function withExtendedResolver(
         return resolve(
           {
             ...context,
-            preferNativePlatform: platform !== 'web',
             resolveRequest: undefined,
-
             mainFields,
 
             // Passing `mainFields` directly won't be considered (in certain version of Metro)